# standard-be/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Standard BE API
  version: 1.0.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PostStatus:
      type: string
      enum: [ DRAFT, PUBLISHED, ARCHIVED ]
...
    ImageRef:
      type: object
      required: [ url ]
      properties:
        url: { type: string }
        alt: { type: string }
    EmbedPlatform:
      type: string
      enum: [ youtube, vimeo ]
    SpotifyType:
      type: string
      enum: [ track, album, playlist, episode, show ]
...
    BlockType:
      type: string
      enum: [ paragraph, heading, image, quote, video, embed, gallery, link, spotify ]
    Block:
      type: object
      required: [ type ]
      properties:
        type:
          $ref: '#/components/schemas/BlockType'
      discriminator:
        propertyName: type
        mapping:
          paragraph: '#/components/schemas/ParagraphBlock'
          heading: '#/components/schemas/HeadingBlock'
          image: '#/components/schemas/ImageBlock'
          quote: '#/components/schemas/QuoteBlock'
          video: '#/components/schemas/VideoBlock'
          embed: '#/components/schemas/EmbedBlock'
          gallery: '#/components/schemas/GalleryBlock'
          link: '#/components/schemas/LinkBlock'
          spotify: '#/components/schemas/SpotifyBlock'
    ParagraphBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ text ]
          properties:
            text:
              type: string
            html:
              type: string
    HeadingBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ level, text ]
          properties:
            level:
              type: integer
              minimum: 1
              maximum: 6
            text:
              type: string
    ImageBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ image ]
          properties:
            image:
              $ref: '#/components/schemas/ImageRef'
    QuoteBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ text ]
          properties:
            text:
              type: string
            author:
              type: string
    VideoBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ url ]
          properties:
            url: { type: string }
            poster: { type: string }
    EmbedBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ platform, id ]
          properties:
            platform: { $ref: '#/components/schemas/EmbedPlatform' }
            id: { type: string }
    GalleryBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ urls ]
          properties:
            urls:
              type: array
              items: { type: string }
    LinkBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ url ]
          properties:
            url: { type: string }
            title: { type: string }
            description: { type: string }
    SpotifyBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ url, spotifyType, spotifyId ]
          properties:
            url: { type: string }
            spotifyType: { $ref: '#/components/schemas/SpotifyType' }
            spotifyId: { type: string }
    PostSummary:
      type: object
      required: [ id, title, slug, status, tags ]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { $ref: '#/components/schemas/PostStatus' }
        thumbnail: { $ref: '#/components/schemas/ImageRef' }
        publishedAt: { type: string, format: date-time }
        date: { type: string, nullable: true }
        duration: { type: string, nullable: true }
    PostDetails:
      type: object
      required: [ id, title, slug, status, tags, content ]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { $ref: '#/components/schemas/PostStatus' }
        thumbnail: { $ref: '#/components/schemas/ImageRef' }
        content:
          type: array
          items: { $ref: '#/components/schemas/Block' }
        publishedAt: { type: string, format: date-time }
        date: { type: string, nullable: true }
        duration: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: string }
...
    PagePostSummary:
      type: object
      required: [ items, page, size ]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PostSummary' }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer }
...
    PodcastImportItem:
      type: object
      required: [ title ]
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string }
        duration: { type: string, nullable: true }
        url: { type: string }
        youtube_id: { type: string, nullable: true }
        youtube_url: { type: string, nullable: true }
        thumbnail: { type: string, nullable: true }
    ImportPostsRequest:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PodcastImportItem' }
    CreatePostRequest:
      type: object
      required: [ title, slug, content ]
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        tags:
          type: array
          items: { type: string }
        thumbnail: { $ref: '#/components/schemas/ImageRef' }
        content:
          type: array
          items: { $ref: '#/components/schemas/Block' }
        date: { type: string, nullable: true }
        duration: { type: string, nullable: true }
    UpdatePostRequest:
      allOf:
        - $ref: '#/components/schemas/CreatePostRequest'
...
paths:
...
  /admin/posts:
    get:
      tags: [ AdminPosts ]
      operationId: adminPostList
      security:
        - bearerAuth: [ ]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/PostStatus' }
      responses:
        "200":
          description: Admin post page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagePostSummary' }
    post:
      tags: [ AdminPosts ]
      operationId: adminPostCreate
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostRequest' }
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostDetails' }
...
  /admin/posts/{id}:
    get:
      tags: [ AdminPosts ]
      operationId: adminPostGet
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostDetails' }
    put:
      tags: [ AdminPosts ]
      operationId: adminPostUpdate
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePostRequest' }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostDetails' }
    delete:
      tags: [ AdminPosts ]
      operationId: adminPostDelete
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
...
  /admin/posts/{id}/publish:
    post:
      tags: [ AdminPosts ]
      operationId: adminPostPublish
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostDetails' }
...
  /admin/posts/import:
    post:
      tags: [ AdminPosts ]
      operationId: adminPostImport
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportPostsRequest' }
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PostSummary' }
...
  /admin/posts/reset:
    post:
      tags: [ AdminPosts ]
      operationId: adminPostReset
      security:
        - bearerAuth: [ ]
      responses:
        "204":
          description: Reset complete
...